server/requirements.txt
# Noxveil - Server Dependencies
# FastAPI web framework
fastapi>=0.109.0

# ASGI server
uvicorn[standard]>=0.27.0

# Async SQLAlchemy for database
sqlalchemy[asyncio]>=2.0.25

# Async SQLite driver
aiosqlite>=0.19.0

# JWT authentication
pyjwt[crypto]>=2.8.0

# Password hashing
bcrypt>=4.1.2

# HTTP client for agent communication
httpx>=0.26.0

# Structured logging
structlog>=24.1.0

# CORS support
python-multipart>=0.0.6

# File uploads
python-magic>=0.4.27

# Date/time utilities
python-dateutil>=2.8.2

# WebSocket support
websockets>=12.0

# Pillow for image processing
Pillow>=10.2.0
